requirements.txt
# -----------------------------
# Core Backend
# -----------------------------
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4

# -----------------------------
# Data Processing
# -----------------------------
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2

# -----------------------------
# Database (SQL for Aggregation)
# -----------------------------
duckdb==0.10.2

# -----------------------------
# Vector DB / RAG
# -----------------------------
chromadb==0.5.0
sentence-transformers==2.7.0

# -----------------------------
# LangChain / LangGraph
# -----------------------------
langchain==0.2.5
langchain-core==0.2.9
langgraph==0.0.55

# -----------------------------
# LLM (Groq)
# -----------------------------
langchain-groq==0.1.4

# -----------------------------
# Frontend (Streamlit)
# -----------------------------
streamlit==1.36.0

# -----------------------------
# HTTP & API Calls
# -----------------------------
requests==2.31.0

# -----------------------------
# Security / OAuth
# -----------------------------
cryptography==41.0.4
python-dotenv==1.0.1

# -----------------------------
# Testing
# -----------------------------
pytest==8.2.2

# -----------------------------
# Optional (Recommended)
# -----------------------------
tqdm==4.66.4